requirements.txt
azure-identity==1.5.0
azure-mgmt-core==1.2.2
azure-mgmt-resource==15.0.0